Replace deprecated addSourceFile calls in AndroidTestMigrate all uses of the deprecatedcom.google.errorprone.CompilationTestHelper#addSourceFile inAndroidTest to addSourceLines with inlined Java text blocks, andremove the android testdata source files. Also drop the class-level@SuppressWarnings("deprecation") annotation, which is no longerneeded.Fixes #1425Signed-off-by: jabrailkhalil <jabrailkhalil@gmail.com>
