+++ nullaway/src/test/java/com/uber/nullaway/AndroidTest.java
@@ -11,8 +11,123 @@ import org.junit.runners.JUnit4;
 
 /** Unit tests for {@link com.uber.nullaway.NullAway}. */
 @RunWith(JUnit4.class)
-@SuppressWarnings("deprecation")
 public class AndroidTest {
+  private static final String CORE_FRAGMENT_STUB =
+      """
+      package android.app;
+
+      import android.content.Context;
+      import android.os.Bundle;
+      import android.view.LayoutInflater;
+      import android.view.View;
+      import android.view.ViewGroup;
+
+      public class Fragment {
+
+        public void onAttach(Context context) {}
+
+        public void onCreate(Bundle savedInstanceState) {}
+
+        public View onCreateView(
+            LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+          return null;
+        }
+      }
+      """;
+
+  private static final String ANDROIDX_FRAGMENT_STUB =
+      """
+      package androidx.fragment.app;
+
+      import android.content.Context;
+      import android.os.Bundle;
+      import android.view.LayoutInflater;
+      import android.view.View;
+      import android.view.ViewGroup;
+
+      public class Fragment {
+
+        public void onAttach(Context context) {}
+
+        public void onCreate(Bundle savedInstanceState) {}
+
+        public View onCreateView(
+            LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+          return null;
+        }
+
+        public void onActivityCreated(Bundle savedInstanceState) {}
+      }
+      """;
+
+  private static final String SUPPORTLIB_FRAGMENT_STUB =
+      """
+      package android.support.v4.app;
+
+      import android.content.Context;
+      import android.os.Bundle;
+      import android.view.LayoutInflater;
+      import android.view.View;
+      import android.view.ViewGroup;
+
+      public class Fragment {
+
+        public void onAttach(Context context) {}
+
+        public void onCreate(Bundle savedInstanceState) {}
+
+        public View onCreateView(
+            LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+          return null;
+        }
+      }
+      """;
+
+  private static final String CORE_ACTIVITY_STUB =
+      """
+      package android.app;
+
+      import android.os.Bundle;
+      import android.os.PersistableBundle;
+
+      public class Activity {
+
+        protected void onCreate(Bundle savedInstanceState) {}
+
+        public void onCreate(Bundle savedInstanceState, PersistableBundle persistentState) {}
+      }
+      """;
+
+  private static final String ANDROIDX_ACTIVITY_COMPAT_STUB =
+      """
+      package androidx.core.app;
+
+      import android.os.Bundle;
+      import android.os.PersistableBundle;
+
+      public class ActivityCompat {
+
+        protected void onCreate(Bundle savedInstanceState) {}
+
+        public void onCreate(Bundle savedInstanceState, PersistableBundle persistentState) {}
+      }
+      """;
+
+  private static final String SUPPORTLIB_ACTIVITY_COMPAT_STUB =
+      """
+      package android.support.v4.app;
+
+      import android.os.Bundle;
+      import android.os.PersistableBundle;
+
+      public class ActivityCompat {
+
+        protected void onCreate(Bundle savedInstanceState) {}
+
+        public void onCreate(Bundle savedInstanceState, PersistableBundle persistentState) {}
+      }
+      """;
+
   @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
 
   private CompilationTestHelper compilationHelper;
@@ -35,8 +150,45 @@ public class AndroidTest {
   public void coreFragmentSuccess() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/core/Fragment.java")
-        .addSourceFile("testdata/android-success/CoreFragment.java")
+        .addSourceLines("androidstubs/core/Fragment.java", CORE_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-success/CoreFragment.java",
+            """
+            package com.uber.myapplication;
+
+            import android.app.Fragment;
+            import android.content.Context;
+            import android.os.Bundle;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+
+            public class CoreFragment extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -44,8 +196,39 @@ public class AndroidTest {
   public void coreFragmentMissingOnAttachError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/core/Fragment.java")
-        .addSourceFile("testdata/android-error/CoreFragmentWithoutOnAttach.java")
+        .addSourceLines("androidstubs/core/Fragment.java", CORE_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/CoreFragmentWithoutOnAttach.java",
+            """
+            package com.uber.myapplication;
+
+            import android.app.Fragment;
+            import android.os.Bundle;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+
+            public class CoreFragmentWithoutOnAttach extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              // BUG: Diagnostic contains: @NonNull field 'mOnAttachInitialisedField' not initialized
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -53,8 +236,39 @@ public class AndroidTest {
   public void coreFragmentMissingOnCreateError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/core/Fragment.java")
-        .addSourceFile("testdata/android-error/CoreFragmentWithoutOnCreate.java")
+        .addSourceLines("androidstubs/core/Fragment.java", CORE_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/CoreFragmentWithoutOnCreate.java",
+            """
+            package com.uber.myapplication;
+
+            import android.app.Fragment;
+            import android.content.Context;
+            import android.os.Bundle;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+
+            public class CoreFragmentWithoutOnCreate extends Fragment {
+              // BUG: Diagnostic contains: @NonNull field 'mOnCreateInitialisedField' not initialized
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -62,8 +276,36 @@ public class AndroidTest {
   public void coreFragmentMissingOnCreateViewError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/core/Fragment.java")
-        .addSourceFile("testdata/android-error/CoreFragmentWithoutOnCreateView.java")
+        .addSourceLines("androidstubs/core/Fragment.java", CORE_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/CoreFragmentWithoutOnCreateView.java",
+            """
+            package com.uber.myapplication;
+
+            import android.app.Fragment;
+            import android.content.Context;
+            import android.os.Bundle;
+
+            public class CoreFragmentWithoutOnCreateView extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              // BUG: Diagnostic contains: @NonNull field 'mOnCreateViewInitialisedField' not initialized
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -72,8 +314,52 @@ public class AndroidTest {
   public void androidxFragmentSuccess() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/androidx/Fragment.java")
-        .addSourceFile("testdata/android-success/AndroidxFragment.java")
+        .addSourceLines("androidstubs/androidx/Fragment.java", ANDROIDX_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-success/AndroidxFragment.java",
+            """
+            package com.uber.myapplication;
+
+            import android.content.Context;
+            import android.os.Bundle;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+            import androidx.fragment.app.Fragment;
+
+            public class AndroidxFragment extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+              private Object monActivityCreatedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+
+              @Override
+              public void onActivityCreated(Bundle savedInstanceState) {
+                super.onActivityCreated(savedInstanceState);
+                monActivityCreatedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -81,8 +367,39 @@ public class AndroidTest {
   public void androidxFragmentMissingOnAttachError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/androidx/Fragment.java")
-        .addSourceFile("testdata/android-error/AndroidxFragmentWithoutOnAttach.java")
+        .addSourceLines("androidstubs/androidx/Fragment.java", ANDROIDX_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/AndroidxFragmentWithoutOnAttach.java",
+            """
+            package com.uber.myapplication;
+
+            import android.os.Bundle;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+            import androidx.fragment.app.Fragment;
+
+            public class AndroidxFragmentWithoutOnAttach extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              // BUG: Diagnostic contains: @NonNull field 'mOnAttachInitialisedField' not initialized
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -90,8 +407,39 @@ public class AndroidTest {
   public void androidxFragmentMissingOnCreateError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/androidx/Fragment.java")
-        .addSourceFile("testdata/android-error/AndroidxFragmentWithoutOnCreate.java")
+        .addSourceLines("androidstubs/androidx/Fragment.java", ANDROIDX_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/AndroidxFragmentWithoutOnCreate.java",
+            """
+            package com.uber.myapplication;
+
+            import android.content.Context;
+            import android.os.Bundle;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+            import androidx.fragment.app.Fragment;
+
+            public class AndroidxFragmentWithoutOnCreate extends Fragment {
+              // BUG: Diagnostic contains: @NonNull field 'mOnCreateInitialisedField' not initialized
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -99,8 +447,36 @@ public class AndroidTest {
   public void androidxFragmentMissingOnCreateViewError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/androidx/Fragment.java")
-        .addSourceFile("testdata/android-error/AndroidxFragmentWithoutOnCreateView.java")
+        .addSourceLines("androidstubs/androidx/Fragment.java", ANDROIDX_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/AndroidxFragmentWithoutOnCreateView.java",
+            """
+            package com.uber.myapplication;
+
+            import android.content.Context;
+            import android.os.Bundle;
+            import androidx.fragment.app.Fragment;
+
+            public class AndroidxFragmentWithoutOnCreateView extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              // BUG: Diagnostic contains: @NonNull field 'mOnCreateViewInitialisedField' not initialized
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -110,8 +486,45 @@ public class AndroidTest {
   public void supportLibFragmentSuccess() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/supportlib/Fragment.java")
-        .addSourceFile("testdata/android-success/SupportLibraryFragment.java")
+        .addSourceLines("androidstubs/supportlib/Fragment.java", SUPPORTLIB_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-success/SupportLibraryFragment.java",
+            """
+            package com.uber.myapplication;
+
+            import android.content.Context;
+            import android.os.Bundle;
+            import android.support.v4.app.Fragment;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+
+            public class SupportLibraryFragment extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -119,8 +532,39 @@ public class AndroidTest {
   public void supportLibFragmentMissingOnAttachError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/supportlib/Fragment.java")
-        .addSourceFile("testdata/android-error/SupportLibraryFragmentWithoutOnAttach.java")
+        .addSourceLines("androidstubs/supportlib/Fragment.java", SUPPORTLIB_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/SupportLibraryFragmentWithoutOnAttach.java",
+            """
+            package com.uber.myapplication;
+
+            import android.os.Bundle;
+            import android.support.v4.app.Fragment;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+
+            public class SupportLibraryFragmentWithoutOnAttach extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              // BUG: Diagnostic contains: @NonNull field 'mOnAttachInitialisedField' not initialized
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -128,8 +572,39 @@ public class AndroidTest {
   public void supportLibFragmentMissingOnCreateError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/supportlib/Fragment.java")
-        .addSourceFile("testdata/android-error/SupportLibraryFragmentWithoutOnCreate.java")
+        .addSourceLines("androidstubs/supportlib/Fragment.java", SUPPORTLIB_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/SupportLibraryFragmentWithoutOnCreate.java",
+            """
+            package com.uber.myapplication;
+
+            import android.content.Context;
+            import android.os.Bundle;
+            import android.support.v4.app.Fragment;
+            import android.view.LayoutInflater;
+            import android.view.View;
+            import android.view.ViewGroup;
+
+            public class SupportLibraryFragmentWithoutOnCreate extends Fragment {
+              // BUG: Diagnostic contains: @NonNull field 'mOnCreateInitialisedField' not initialized
+              private Object mOnCreateInitialisedField;
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public View onCreateView(
+                  LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+                mOnCreateViewInitialisedField = new Object();
+                return super.onCreateView(inflater, container, savedInstanceState);
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -137,8 +612,36 @@ public class AndroidTest {
   public void supportLibFragmentMissingOnCreateViewError() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/supportlib/Fragment.java")
-        .addSourceFile("testdata/android-error/SupportLibraryFragmentWithoutOnCreateView.java")
+        .addSourceLines("androidstubs/supportlib/Fragment.java", SUPPORTLIB_FRAGMENT_STUB)
+        .addSourceLines(
+            "android-error/SupportLibraryFragmentWithoutOnCreateView.java",
+            """
+            package com.uber.myapplication;
+
+            import android.content.Context;
+            import android.os.Bundle;
+            import android.support.v4.app.Fragment;
+
+            public class SupportLibraryFragmentWithoutOnCreateView extends Fragment {
+
+              private Object mOnCreateInitialisedField;
+              // BUG: Diagnostic contains: @NonNull field 'mOnCreateViewInitialisedField' not initialized
+              private Object mOnCreateViewInitialisedField;
+              private Object mOnAttachInitialisedField;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField = new Object();
+              }
+
+              @Override
+              public void onAttach(Context context) {
+                super.onAttach(context);
+                mOnAttachInitialisedField = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -148,8 +651,34 @@ public class AndroidTest {
   public void coreActivitySuccess() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/core/Activity.java")
-        .addSourceFile("testdata/android-success/CoreActivity.java")
+        .addSourceLines("androidstubs/core/Activity.java", CORE_ACTIVITY_STUB)
+        .addSourceLines(
+            "android-success/CoreActivity.java",
+            """
+            package com.uber.myapplication;
+
+            import android.app.Activity;
+            import android.os.Bundle;
+            import android.os.PersistableBundle;
+
+            public class CoreActivity extends Activity {
+
+              private Object mOnCreateInitialisedField1;
+              private Object mOnCreateInitialisedField2;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField1 = new Object();
+              }
+
+              @Override
+              public void onCreate(Bundle savedInstanceState, PersistableBundle persistentState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField2 = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -159,8 +688,35 @@ public class AndroidTest {
   public void supportLibActivitySuccess() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/supportlib/ActivityCompat.java")
-        .addSourceFile("testdata/android-success/SupportLibActivityCompat.java")
+        .addSourceLines(
+            "androidstubs/supportlib/ActivityCompat.java", SUPPORTLIB_ACTIVITY_COMPAT_STUB)
+        .addSourceLines(
+            "android-success/SupportLibActivityCompat.java",
+            """
+            package com.uber.myapplication;
+
+            import android.os.Bundle;
+            import android.os.PersistableBundle;
+            import android.support.v4.app.ActivityCompat;
+
+            public class SupportLibActivityCompat extends ActivityCompat {
+
+              private Object mOnCreateInitialisedField1;
+              private Object mOnCreateInitialisedField2;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField1 = new Object();
+              }
+
+              @Override
+              public void onCreate(Bundle savedInstanceState, PersistableBundle persistentState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField2 = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -170,8 +726,34 @@ public class AndroidTest {
   public void androidxActivitySuccess() {
     initializeAndroidCoreClasses();
     compilationHelper
-        .addSourceFile("testdata/androidstubs/androidx/ActivityCompat.java")
-        .addSourceFile("testdata/android-success/AndroidxActivityCompat.java")
+        .addSourceLines("androidstubs/androidx/ActivityCompat.java", ANDROIDX_ACTIVITY_COMPAT_STUB)
+        .addSourceLines(
+            "android-success/AndroidxActivityCompat.java",
+            """
+            package com.uber.myapplication;
+
+            import android.os.Bundle;
+            import android.os.PersistableBundle;
+            import androidx.core.app.ActivityCompat;
+
+            public class AndroidxActivityCompat extends ActivityCompat {
+
+              private Object mOnCreateInitialisedField1;
+              private Object mOnCreateInitialisedField2;
+
+              @Override
+              public void onCreate(Bundle savedInstanceState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField1 = new Object();
+              }
+
+              @Override
+              public void onCreate(Bundle savedInstanceState, PersistableBundle persistentState) {
+                super.onCreate(savedInstanceState);
+                mOnCreateInitialisedField2 = new Object();
+              }
+            }
+            """)
         .doTest();
   }
 
@@ -179,11 +761,47 @@ public class AndroidTest {
   @SuppressWarnings("CheckReturnValue")
   private void initializeAndroidCoreClasses() {
     compilationHelper
-        .addSourceFile("testdata/androidstubs/core/Context.java")
-        .addSourceFile("testdata/androidstubs/core/Bundle.java")
-        .addSourceFile("testdata/androidstubs/core/LayoutInflater.java")
-        .addSourceFile("testdata/androidstubs/core/PersistableBundle.java")
-        .addSourceFile("testdata/androidstubs/core/View.java")
-        .addSourceFile("testdata/androidstubs/core/ViewGroup.java");
+        .addSourceLines(
+            "androidstubs/core/Context.java",
+            """
+            package android.content;
+
+            public class Context {}
+            """)
+        .addSourceLines(
+            "androidstubs/core/Bundle.java",
+            """
+            package android.os;
+
+            public class Bundle {}
+            """)
+        .addSourceLines(
+            "androidstubs/core/LayoutInflater.java",
+            """
+            package android.view;
+
+            public class LayoutInflater {}
+            """)
+        .addSourceLines(
+            "androidstubs/core/PersistableBundle.java",
+            """
+            package android.os;
+
+            public class PersistableBundle {}
+            """)
+        .addSourceLines(
+            "androidstubs/core/View.java",
+            """
+            package android.view;
+
+            public class View {}
+            """)
+        .addSourceLines(
+            "androidstubs/core/ViewGroup.java",
+            """
+            package android.view;
+
+            public class ViewGroup {}
+            """);
   }
 }
